heavily refactored frontend ct
